应用管理模块支持群控

+++ XModule/Xky.XModule.AppManager/ModulePanel.xaml.cs
@@ -31,7 +31,7 @@ namespace Xky.XModule.AppManager
         {
             InitializeComponent();
         }
-        public Device device;
+        public List<Device> devices;
         public string CurrentDirectory = "/";
         public ObservableCollection<DeviceApp> DeviceApps = new ObservableCollection<DeviceApp>();
         public Dictionary<string, string> PackageNames = new Dictionary<string, string>();
@@ -45,21 +45,21 @@ namespace Xky.XModule.AppManager
                 Stream myStream = myAssembly.GetManifestResourceStream("Xky.XModule.AppManager.packages.txt");
                 byte[] bytes = new byte[myStream.Length];
                 myStream.Read(bytes, 0, bytes.Length);
-               string []infos= Encoding.UTF8.GetString(bytes).Split('\n');
+                string[] infos = Encoding.UTF8.GetString(bytes).Split('\n');
                 foreach (string info in infos)
                 {
-                    string inf= info.Trim();
-                    if (inf.Length>0&&inf.Contains("|"))
+                    string inf = info.Trim();
+                    if (inf.Length > 0 && inf.Contains("|"))
                     {
-                        string []infs = inf.Split('|');
+                        string[] infs = inf.Split('|');
                         string appname = infs[0];
                         string packagename = infs[1];
                         if (!PackageNames.ContainsKey(packagename))
                         {
-                            PackageNames.Add(packagename,appname);
+                            PackageNames.Add(packagename, appname);
                         }
                     }
-           
+
 
                 }
                 ShowLoading("正在加载设备应用列表...");
@@ -68,7 +68,7 @@ namespace Xky.XModule.AppManager
             })
             { IsBackground = true }.Start();
         }
- 
+
         public class DeviceApp
         {
             private string _name = "";
@@ -79,79 +79,81 @@ namespace Xky.XModule.AppManager
             public string PackageName { get => _packageName; set => _packageName = value; }
         }
 
-       
+
         public void LoadPackages()
         {
-    
-           
-                List<DeviceApp> list = new List<DeviceApp>();
-                Response res_system = device.ScriptEngine.AdbShell("pm list package -s");
-                if (res_system.Json["result"] != null)
+
+
+            List<DeviceApp> list = new List<DeviceApp>();
+            Response res_system = devices[0].ScriptEngine.AdbShell("pm list package -s");
+            if (res_system.Json["result"] != null)
+            {
+                List<string> res = res_system.Json["result"].ToString().Split('\n').ToList();
+                foreach (string s in res)
                 {
-                    List<string> res = res_system.Json["result"].ToString().Split('\n').ToList();
-                    foreach (string s in res)
+                    if (s.StartsWith("package:"))
                     {
-                        if (s.StartsWith("package:"))
+                        DeviceApp deviceApp = new DeviceApp();
+                        deviceApp.Type = "system";
+                        int index = s.IndexOf("package:") + 8;
+                        deviceApp.PackageName = s.Substring(index, s.Length - index);
+                        if (PackageNames.ContainsKey(deviceApp.PackageName))
                         {
-                            DeviceApp deviceApp = new DeviceApp();
-                            deviceApp.Type = "system";
-                            int index = s.IndexOf("package:") + 8;
-                            deviceApp.PackageName = s.Substring(index, s.Length - index);
-                            if (PackageNames.ContainsKey(deviceApp.PackageName))
-                            {
-                                deviceApp.Name = PackageNames[deviceApp.PackageName];
-                            }
-                            else {
-                                deviceApp.Name = deviceApp.PackageName;
-                            }
-                            list.Add(deviceApp);
+                            deviceApp.Name = PackageNames[deviceApp.PackageName];
                         }
-                    }
-
-                }
-                res_system = device.ScriptEngine.AdbShell("pm list package -3");
-                if (res_system.Json["result"] != null)
-                {
-                    List<string> res = res_system.Json["result"].ToString().Split('\n').ToList();
-                    foreach (string s in res)
-                    {
-                        if (s.StartsWith("package:"))
+                        else
                         {
-                            DeviceApp deviceApp = new DeviceApp();
-                            deviceApp.Type = "user";
-                            int index = s.IndexOf("package:") + 8;
-                            deviceApp.PackageName = s.Substring(index, s.Length - index);
-                            if (PackageNames.ContainsKey(deviceApp.PackageName))
-                            {
-                                deviceApp.Name = PackageNames[deviceApp.PackageName];
-                            }
-                            else
-                            {
-                                deviceApp.Name = deviceApp.PackageName;
-                            }
-                            list.Add(deviceApp);
+                            deviceApp.Name = deviceApp.PackageName;
                         }
+                        list.Add(deviceApp);
                     }
-
                 }
-                list.Sort((left, right) =>
-                {
-                    if (left.Type != right.Type)
-                    {
-                        if (left.Type == "system") { return 1; }
-                        else { return -1; }
-                    }
-                    else { return 0; }
-                });
-                DeviceApps = new ObservableCollection<DeviceApp>(list);
-                this.Dispatcher.Invoke(new Action(() =>
-                {
-                    ItemListBox.ItemsSource = DeviceApps;
-                }));
 
-            
+            }
+            res_system = devices[0].ScriptEngine.AdbShell("pm list package -3");
+            if (res_system.Json["result"] != null)
+            {
+                List<string> res = res_system.Json["result"].ToString().Split('\n').ToList();
+                foreach (string s in res)
+                {
+                    if (s.StartsWith("package:"))
+                    {
+                        DeviceApp deviceApp = new DeviceApp();
+                        deviceApp.Type = "user";
+                        int index = s.IndexOf("package:") + 8;
+                        deviceApp.PackageName = s.Substring(index, s.Length - index);
+                        if (PackageNames.ContainsKey(deviceApp.PackageName))
+                        {
+                            deviceApp.Name = PackageNames[deviceApp.PackageName];
+                        }
+                        else
+                        {
+                            deviceApp.Name = deviceApp.PackageName;
+                        }
+                        list.Add(deviceApp);
+                    }
+                }
+
+            }
+            list.Sort((left, right) =>
+            {
+                if (left.Type != right.Type)
+                {
+                    if (left.Type == "system") { return 1; }
+                    else { return -1; }
+                }
+                else { return 0; }
+            });
+            DeviceApps = new ObservableCollection<DeviceApp>(list);
+            this.Dispatcher.Invoke(new Action(() =>
+            {
+                ItemListBox.ItemsSource = DeviceApps;
+            }));
+
+
         }
-        public void ReLoadPackages() {
+        public void ReLoadPackages()
+        {
             new Thread(() =>
             {
                 ShowLoading("正在加载设备应用列表...");
@@ -168,20 +170,27 @@ namespace Xky.XModule.AppManager
                 foreach (var obj in arr)
                 {
                     string filename = obj.ToString();
-                    device.ScriptEngine.WriteBufferToFile(CurrentDirectory+"/"+new FileInfo(filename).Name, File.ReadAllBytes(filename));
-                    Console.WriteLine("文件上传完毕");
+                    for (int i = 0; i < devices.Count; i++)
+                    {
+                        if (devices.Count > i)
+                        {
+                            devices[i].ScriptEngine.WriteBufferToFile(CurrentDirectory + "/" + new FileInfo(filename).Name, File.ReadAllBytes(filename));
+                            Console.WriteLine("文件上传完毕");
+                        }
+                    }
+
                 }
-               
+
             }
 
-            
+
         }
         private void MenuItem_Open_Click(object sender, RoutedEventArgs e)
         {
             //DeviceApp deviceApp = (DeviceApp)((System.Windows.Controls.MenuItem)sender).DataContext;
             //Console.WriteLine(deviceApp.PackageName);
             //Response res = device.ScriptEngine.RestartApp(deviceApp.PackageName);
-          
+
         }
         private void MenuItem_Stop_Click(object sender, RoutedEventArgs e)
         {
@@ -191,7 +200,7 @@ namespace Xky.XModule.AppManager
         {
 
         }
-        
+
         private void MenuItem_Install_Click(object sender, RoutedEventArgs e)
         {
 
@@ -247,48 +256,56 @@ namespace Xky.XModule.AppManager
             if (filename_new.Length > 50)
             {
                 filename_new.Remove(50);
-            }
-            if (fi.Length > 10485760)
+            }//大于10M的文件采用分段上传
+            for (int index = 0; index < devices.Count; index++)
             {
-                //大于10M的文件采用分段上传
-              
-                FileStream fs = new FileStream(filename, FileMode.Open);
-               
-                string tempdir = filename_new + "_temp";
-                int i = 1;
-                byte[] bytes = new byte[10485760];
-                string catf = "";
-                int readlength = fs.Read(bytes, 0, 10485760);
-                while (readlength > 0)
+                if (devices.Count > index)
                 {
-                    ShowLoading("使用大文件上传模式，预计还需" + ((fi.Length / 10485760)-i+1)+ "秒钟...");
-                    if (readlength < 10485760)
+                    var device = devices[index];
+                    ShowLoading("正在上传文件[" + filename + "]到设备["+device.Name+"]...");
+                    if (fi.Length > 10485760)
                     {
-                        device.ScriptEngine.WriteBufferToFile(dir + "/" + tempdir + "/" + i + ".tmp", bytes.Take(readlength).ToArray());
 
-                    }
-                    else { device.ScriptEngine.WriteBufferToFile(dir + "/" + tempdir + "/" + i + ".tmp", bytes); }
-                    Console.WriteLine("正在上传第" + i + "个拆分文件，大小" + readlength + "...");
-                    catf += i + ".tmp  ";
+                        FileStream fs = new FileStream(filename, FileMode.Open);
 
-                    i++;
-                    readlength = fs.Read(bytes, 0, 10485760);
-                  
+                        string tempdir = filename_new + "_temp";
+                        int i = 1;
+                        byte[] bytes = new byte[10485760];
+                        string catf = "";
+                        int readlength = fs.Read(bytes, 0, 10485760);
+                        while (readlength > 0)
+                        {
+                            ShowLoading("设备["+device.Name+"]使用大文件上传模式，预计还需" + ((fi.Length / 10485760) - i + 1) + "秒钟...");
+                            if (readlength < 10485760)
+                            {
+                                device.ScriptEngine.WriteBufferToFile(dir + "/" + tempdir + "/" + i + ".tmp", bytes.Take(readlength).ToArray());
+
+                            }
+                            else { device.ScriptEngine.WriteBufferToFile(dir + "/" + tempdir + "/" + i + ".tmp", bytes); }
+                            Console.WriteLine("正在上传第" + i + "个拆分文件，大小" + readlength + "...");
+                            catf += i + ".tmp  ";
+
+                            i++;
+                            readlength = fs.Read(bytes, 0, 10485760);
+
+                        }
+                        fs.Close();
+                        for (int w = i / 2; w > 0; w--)
+                        {
+                            ShowLoading("设备["+device.Name+"]正在合并文件，预计还需" + w + "秒钟...");
+                            Thread.Sleep(1000);
+                        }
+
+                        Response res = device.ScriptEngine.AdbShell("cd " + dir + "/" + tempdir + "&&cat " + catf + " > ../" + filename_new + fi.Extension + "&&rm -r -f " + dir + "/" + tempdir);
+                    
                 }
-                fs.Close();
-                for (int w = i / 2; w >0; w--)
+                else
                 {
-                    ShowLoading("正在合并文件，预计还需" + w + "秒钟...");
-                    Thread.Sleep(1000);
+                    Response res = device.ScriptEngine.WriteBufferToFile(dir + "/" + new FileInfo(filename).Name, File.ReadAllBytes(filename));
                 }
-               
-                Response res = device.ScriptEngine.AdbShell("cd " + dir + "/" + tempdir + "&&cat " + catf + " > ../" + filename_new + fi.Extension + "&&rm -r -f " + dir + "/" + tempdir);
-
             }
-            else
-            {
-                Response res = device.ScriptEngine.WriteBufferToFile(dir + "/" + new FileInfo(filename).Name, File.ReadAllBytes(filename));
-            }
+        }
+    
 
         }
         private void Button_Close_Click(object sender, RoutedEventArgs e)
@@ -313,8 +330,17 @@ namespace Xky.XModule.AppManager
                     string filename = openFileDialog.FileName.ToString();
                     UploadFile("/sdcard", filename);
               
-                    ShowLoading("正在安装应用...");
-                    Response res = device.ScriptEngine.AdbShell("pm install -g -r /sdcard/" + new FileInfo(filename).Name+"&");
+                   
+                    for (int index = 0; index < devices.Count; index++)
+                    {
+                        if (devices.Count > index)
+                        {
+                            var device = devices[index];
+                            ShowLoading("正在安装应用到设备["+device.Name+"]...");
+                            Response res = device.ScriptEngine.AdbShell("pm install -g -r -i -d /sdcard/" + new FileInfo(filename).Name + "&");
+                            Console.WriteLine("设备["+device.Name+"]安装结果："+res.Json.ToString());
+                        }
+                    }
                     Thread.Sleep(5000);
                     ShowLoading("成功发送安装命令，请稍后刷新应用列表查看是否成功安装！");
                     Thread.Sleep(5000);
